requirements.txt
Flask
python-dotenv
google-cloud-bigquery
google-cloud-monitoring
google-cloud-recommender
google-cloud-firestore
google.generativeai
google-cloud-aiplatform
google-cloud-compute
